fix(script.sh): save year and author in LICENSE file
Added an option to pass author name from the cli which will be placed in LICENSE file. If author is
not passed the script will default to using the git_user passed with -u option or read from git
global config. Where appropritae it will replace the placeholders in the LICENSE file with author
and year. (bug fix) It will instead use 'git add .' when staging files.